requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
python-dotenv==1.0.0
cryptography==41.0.3
Werkzeug==2.3.6
requests==2.31.0
email-validator==2.0.0
redis==4.6.0
bleach==6.0.0
